extract-text/requirements.txt
pdfminer.six
python-pptx